Add styleguide, start prototype
This is the last version since projet should move to a Django.

Peewee and CherryPy does not work correctly together naturally, it needs
a little more work to achieve this because of some problem with threads, CherryPy
working in its own.

Also modelization is not very finished since it lack of folder tree features
and discovery may be changed also. Peewee is not as natural as Django ORM so
it does not help.

And we may think about using meta tags to have more infos on videos.

Finally, frontend has not really started because problem have been discovered
during prototyping.